feat & fix: add card feature & fix last row in cards (again)
add the ability to create your own cards

fix the alignment of the last row of cards
Earlier, I did not take into account the fact that the card
may lack two or one. That is, the situation was previously
corrected for only one case. Still left the flexbox type for the parent